+++ docs/versionhistory.rst
@@ -14,8 +14,10 @@ Version history
   ``asphalt.core.util`` module
 - *BACKWARD INCOMPATIBLE* Removed regular context manager support from the ``Context`` class
   (asynchronous context manager support still remains)
+- *BACKWARD INCOMPATIBLE* Modified event dispatch logic in ``EventSource`` to always run all
+  event listeners even if some listeners raise exceptions. A uniform exception is then raised
+  that contains all the exceptions and the listeners who raised them.
 - Added the ability to listen to multiple topics in an EventSource with a single listener
-
 
 **1.3.0**